add methods to save and load custom names from disk
This helps preserve custom names across restarts.

Add functions to save and load a dictionary of workspace numbers
to custom names. On startup, read these saved names from file to
populate the current tree with custom names. Whenever a custom name
is added or removed, the dictionary is pickled and written to file.

If save_file is None, then turn off this functionality.

* add `save_file` element to the schema
* add functions `save_names` and `read_saved_names`
* add utility function `get_workspace_by_num`